Refactored code
Code was refactored in a way more appropiate with industry standard
