+++ api/middleware/tokenGenerator.js
